// __server/domain/schemas/Symbol.schema.ts
import { z } from 'zod';

export const SymbolSchema = z.object({
  ask: z.number(),
  bid: z.number(),
  categoryName: z.string(),
  contractSize: z.number(),
  currency: z.string(),
  currencyPair: z.boolean(),
  currencyProfit: z.string(),
  description: z.string(),
  expiration: z.union([z.string(), z.null()]),
  groupName: z.string(),
  high: z.number(),
  initialMargin: z.number(),
  instantMaxVolume: z.number(),
  leverage: z.number(),
  longOnly: z.boolean(),
  lotMax: z.number(),
  lotMin: z.number(),
  lotStep: z.number(),
  low: z.number(),
  marginHedged: z.number(),
  marginHedgedStrong: z.boolean(),
  marginMaintenance: z.union([z.number(), z.null()]),
  marginMode: z.number(),
  percentage: z.number(),
  pipsPrecision: z.number(),
  precision: z.number(),
  profitMode: z.number(),
  quoteId: z.number(),
  shortSelling: z.boolean(),
  spreadRaw: z.number(),
  spreadTable: z.number(),
  starting: z.union([z.string(), z.null()]),
  stepRuleId: z.number(),
  stopsLevel: z.number(),
  swap_rollover3days: z.number(),
  swapEnable: z.boolean(),
  swapLong: z.number(),
  swapShort: z.number(),
  swapType: z.number(),
  symbol: z.string(),
  tickSize: z.union([z.number(), z.null()]),
  tickValue: z.union([z.number(), z.null()]),
  time: z.number(),
  timeString: z.string(),
  trailingEnabled: z.boolean(),
  type: z.number(),
});

// [0]     "symbol": "EIMI.UK",
// [0]     "currency": "USD",
// [0]     "categoryName": "ETF",
// [0]     "currencyProfit": "USD",
// [0]     "quoteId": 6,
// [0]     "quoteIdCross": 15,
// [0]     "marginMode": 104,
// [0]     "profitMode": 6,
// [0]     "pipsPrecision": 3,
// [0]     "contractSize": 1,
// [0]     "exemode": 1,
// [0]     "time": 1739550519926,
// [0]     "expiration": null,
// [0]     "stopsLevel": 0,
// [0]     "precision": 3,
// [0]     "swapType": 2,
// [0]     "stepRuleId": 323,
// [0]     "type": 3375,
// [0]     "instantMaxVolume": 2147483647,
// [0]     "groupName": "ETFs",
// [0]     "description": "iShares, UCITS, ACC, USD",
// [0]     "longOnly": true,
// [0]     "trailingEnabled": false,
// [0]     "marginHedgedStrong": false,
// [0]     "swapEnable": true,
// [0]     "percentage": 100,
// [0]     "bid": 35.41,
// [0]     "ask": 35.5,
// [0]     "high": 35.51,
// [0]     "low": 35.26,
// [0]     "lotMin": 1,
// [0]     "lotMax": 1000000,
// [0]     "lotStep": 1,
// [0]     "tickSize": 0.001,
// [0]     "tickValue": 0.001,
// [0]     "swapLong": 0,
// [0]     "swapShort": 0,
// [0]     "leverage": 100,
// [0]     "spreadRaw": 0.09,
// [0]     "spreadTable": 90,
// [0]     "starting": null,
// [0]     "swap_rollover3days": 0,
// [0]     "marginMaintenance": 0,
// [0]     "marginHedged": 0,
// [0]     "initialMargin": 0,
// [0]     "timeString": "Fri Feb 14 17:28:39 CET 2025",
// [0]     "shortSelling": false,
// [0]     "currencyPair": false

export const SymbolResponseSchema = z.object({
  status: z.boolean(),
  returnData: SymbolSchema,
});

export type SymbolType = z.infer<typeof SymbolSchema>;
export type SymbolResponseType = z.infer<typeof SymbolResponseSchema>;